requirements.txt
flask==3.0.0
anthropic==0.8.1
python-dotenv==1.0.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3 